Add exercise and set components with functionality for managing workouts
