+++ frontend/src/app/features/workouts/components/set/set.component.scss
@@ -1,0 +1,124 @@
+.sets {
+  padding: 0;
+  list-style: none;
+
+  button {
+    padding: 0.5rem 1rem;
+    background-color: #007bff;
+    color: #fff;
+    border: none;
+    border-radius: 0.25rem;
+    cursor: pointer;
+    transition: background-color 0.2s;
+
+    &:hover {
+      background-color: #0056b3;
+    }
+
+    &:disabled {
+      background-color: #ccc;
+      cursor: not-allowed;
+    }
+  }
+
+  .set {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    padding: 0.5rem 0.75rem;
+    border-bottom: 1px solid #f0f0f0;
+
+    .set-info {
+      display: flex;
+
+      p {
+        margin: 0;
+        font-size: 1rem;
+        color: #333;
+      }
+
+      .delete-set {
+        cursor: pointer;
+        color: #e63946;
+        margin-left: 15px;
+        border: none;
+        font-size: 1.2rem;
+        font-weight: bold;
+        transition: color 0.2s;
+
+        &:hover {
+          color: #d62839;
+        }
+      }
+    }
+
+    .set-actions {
+      button {
+        margin-left: 30px;
+      }
+    }
+
+    &:last-child {
+      border-bottom: none;
+    }
+
+    &.completed {
+      background-color: #e6f7e6;
+      color: #2d6a4f;
+
+      p {
+        text-decoration: line-through;
+      }
+    }
+
+    span {
+      color: #28a745;
+      font-weight: bold;
+      cursor: pointer;
+      transition: color 0.2s;
+
+      &:hover {
+        color: #218838;
+      }
+    }
+  }
+}
+
+.add-set {
+  display: flex;
+  gap: 0.5rem;
+  margin-top: 1rem;
+
+  button {
+    padding: 0.5rem 1rem;
+    background-color: #007bff;
+    color: #fff;
+    border: none;
+    border-radius: 0.25rem;
+    cursor: pointer;
+    transition: background-color 0.2s;
+
+    &:hover {
+      background-color: #0056b3;
+    }
+
+    &:disabled {
+      background-color: #ccc;
+      cursor: not-allowed;
+    }
+  }
+
+  input {
+    width: 80px;
+    padding: 0.4rem;
+    border: 1px solid #ccc;
+    border-radius: 0.25rem;
+    text-align: center;
+
+    &:focus {
+      border-color: #007bff;
+      outline: none;
+      box-shadow: 0 0 3px rgba(0, 123, 255, 0.5);
+    }
+  }
+}
