+++ frontend/src/app/features/workouts/components/workout/workout.component.scss
@@ -46,146 +46,6 @@
     }
   }
 
-  .add-exercise {
-    input {
-      width: 100%;
-      padding: 0.5rem;
-      margin-bottom: 1rem;
-      border: 1px solid #ccc;
-      border-radius: 0.5rem;
-      font-weight: bold;
-      font-size: 1rem;
-      box-sizing: border-box;
-
-      &:focus {
-        border-color: #007bff;
-        outline: none;
-        box-shadow: 0 0 3px rgba(0, 123, 255, 0.5);
-      }
-    }
-  }
-
-  .exercise {
-    margin: 1.5rem 0;
-    padding: 1rem 1rem 2rem 1rem;
-    background: #fff;
-    border-radius: 0.75rem;
-    box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
-
-    .exercise-header {
-      display: flex;
-      justify-content: space-between;
-      align-items: center;
-      margin-bottom: 1rem;
-
-      h2 {
-        font-size: 1.2rem;
-        color: #333;
-        margin: 0;
-      }
-
-      .delete-exercise {
-        font-size: 1.5rem;
-        font-weight: bold;
-        border: none;
-        color: #e63946;
-        cursor: pointer;
-        transition: color 0.2s;
-
-        &:hover {
-          color: #d62839;
-        }
-      }
-    }
-
-    .sets {
-      padding: 0;
-      list-style: none;
-
-      .set {
-        display: flex;
-        align-items: center;
-        justify-content: space-between;
-        padding: 0.5rem 0.75rem;
-        border-bottom: 1px solid #f0f0f0;
-
-        .set-info {
-          display: flex;
-
-          p {
-            margin: 0;
-            font-size: 1rem;
-            color: #333;
-          }
-
-          .delete-set {
-            cursor: pointer;
-            color: #e63946;
-            margin-left: 15px;
-            border: none;
-            font-size: 1.2rem;
-            font-weight: bold;
-            transition: color 0.2s;
-
-            &:hover {
-              color: #d62839;
-            }
-          }
-        }
-
-        .set-actions {
-          button {
-            margin-left: 30px;
-          }
-        }
-
-        &:last-child {
-          border-bottom: none;
-        }
-
-        &.completed {
-          background-color: #e6f7e6;
-          color: #2d6a4f;
-
-          p {
-            text-decoration: line-through;
-          }
-        }
-
-        span {
-          color: #28a745;
-          font-weight: bold;
-          cursor: pointer;
-          transition: color 0.2s;
-
-          &:hover {
-            color: #218838;
-          }
-        }
-      }
-    }
-
-    .add-set {
-      display: flex;
-      gap: 0.5rem;
-      margin-top: 1rem;
-
-      input {
-        width: 80px;
-        padding: 0.4rem;
-        border: 1px solid #ccc;
-        border-radius: 0.25rem;
-        text-align: center;
-
-        &:focus {
-          border-color: #007bff;
-          outline: none;
-          box-shadow: 0 0 3px rgba(0, 123, 255, 0.5);
-        }
-      }
-    }
-  }
-
   .workout-completed {
     margin-top: 2rem;
     text-align: center;
